fix: connecting to session with a custom name
The current behavior of sesh is to assume that session names match the session
path. Which causes new session to be created when attempting to connect to a
session where the name does not match the path. This changes that behavior to
preserve name of tmux sessions when connecting to existing sessions.