refactor: tmux command execution
Restructuring the tmux package to use a central struct for command execution to help
with testability.

To limit the scope of the this PR it includes the use of a package level instance of
the new tmux.Command struct and an init function to initialize it. Once all of the
functions using the tmux cli are updated to use the Command struct directly that
code can be removed in favor of a single tmux.Command instance configured in the cli.