refactor: preserve src order by score offset
In the original implementation using for loops, the order of the
sessions relied on the order precreated by the source strategies.

This gets brittle, if we want to remove duplicates from the whole
session list with some internal ranking using the higher abstractions
from the lo library.

Because we already have a field for ranking from the zoxide feature, we
can use it for the other sources as well.
The individual score is created by a source-specific offset and the
original position in the source-specific list.